Resolved issue - cleaned up redundancy
TL;DR - add the database name ("matchiq" without quotes) to the end of your .env files

This branch is meant to resolve the issue of not being able to see the jobs collection data when going to localhost:4000/api/jobs - this issue was due to the mongo_utils.py file specifying a connection to the matchiq database, while nothing else in our code was specifically pointing to that same database. The variable used to connecto to our MongoDB database, DB_URI, was pointing to our MongoDB account without specifying a database, which meant it was defaulting to a database called "test". This is also why the new user registration page was creating users under the "test" database. Appending "matchiq" to the end of our .env files will resolve this issue. I removed the line that defines the db since it will be redundant after the .env file change.